# assets/openapi.json
{
    "openapi": "3.0.0",
    "servers": [
        {
            "url": "http://localhost:3001/api/",
            "description": "Official Solarcord Instance"
        }
    ],
    "info": {
        "description": "Solarcord is a free open source selfhostable discord compatible chat, voice and video platform",
        "version": "1.0.0",
        "title": "Solarcord HTTP API Routes",
        "termsOfService": "",
        "contact": {
            "name": "Solarcord"
        },
        "license": {
            "name": "AGPLV3",
            "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer",
                "description": "Bearer/Bot prefixes are not required."
            }
        },
        "schemas": {
...
            "ChannelPermissionOverwriteType": {
                "enum": [
                    0,
                    1,
                    2
                ],
                "type": "number"
            },
            "Embed": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "type": {
                        "enum": [
                            "article",
                            "gifv",
                            "image",
                            "link",
                            "rich",
                            "video"
                        ],
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "url": {
                        "type": "string"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "color": {
                        "type": "integer"
                    },
                    "footer": {
                        "type": "object",
                        "properties": {
                            "text": {
                                "type": "string"
                            },
                            "icon_url": {
                                "type": "string"
                            },
                            "proxy_icon_url": {
                                "type": "string"
                            }
                        },
                        "additionalProperties": false,
                        "required": [
                            "text"
                        ]
                    },
                    "image": {
                        "$ref": "#/components/schemas/EmbedImage"
                    },
                    "thumbnail": {
                        "$ref": "#/components/schemas/EmbedImage"
                    },
                    "video": {
                        "$ref": "#/components/schemas/EmbedImage"
                    },
                    "provider": {
                        "type": "object",
                        "properties": {
                            "name": {
                                "type": "string"
                            },
                            "url": {
                                "type": "string"
                            }
                        },
                        "additionalProperties": false
                    },
                    "author": {
                        "type": "object",
                        "properties": {
                            "name": {
                                "type": "string"
                            },
                            "url": {
                                "type": "string"
                            },
                            "icon_url": {
                                "type": "string"
                            },
                            "proxy_icon_url": {
                                "type": "string"
                            }
                        },
                        "additionalProperties": false
                    },
                    "fields": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "value": {
                                    "type": "string"
                                },
                                "inline": {
                                    "type": "boolean"
                                }
                            },
                            "additionalProperties": false,
                            "required": [
                                "name",
                                "value"
                            ]
                        }
                    }
                }
            },
            "EmbedImage": {
                "type": "object",
                "properties": {
                    "url": {
                        "type": "string"
                    },
                    "proxy_url": {
                        "type": "string"
                    },
                    "height": {
                        "type": "integer"
                    },
                    "width": {
                        "type": "integer"
                    }
                }
            },
            "ChannelModifySchema": {
                "type": "object",
                "properties": {
                    "name": {
                        "maxLength": 100,
                        "type": "string"
                    },
                    "type": {
                        "enum": [
                            0,
                            1,
                            10,
                            11,
                            12,
                            13,
                            14,
                            15,
                            2,
                            255,
                            3,
                            33,
                            34,
                            35,
                            4,
                            5,
                            6,
                            64,
                            7,
                            8,
                            9
                        ],
                        "type": "number"
                    },
                    "topic": {
                        "type": "string"
                    },
                    "icon": {
                        "type": "string",
                        "nullable": true
                    },
                    "bitrate": {
                        "type": "integer"
                    },
                    "user_limit": {
                        "type": "integer"
                    },
                    "rate_limit_per_user": {
                        "type": "integer"
                    },
                    "position": {
                        "type": "integer"
                    },
                    "permission_overwrites": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "type": {
                                    "$ref": "#/components/schemas/ChannelPermissionOverwriteType"
                                },
                                "allow": {
                                    "type": "string"
                                },
                                "deny": {
                                    "type": "string"
                                }
                            },
                            "additionalProperties": false,
                            "required": [
                                "allow",
                                "deny",
                                "id",
                                "type"
                            ]
                        }
                    },
                    "parent_id": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "nsfw": {
                        "type": "boolean"
                    },
                    "rtc_region": {
                        "type": "string"
                    },
                    "default_auto_archive_duration": {
                        "type": "integer"
                    },
                    "default_reaction_emoji": {
                        "type": "string",
                        "nullable": true
                    },
                    "flags": {
                        "type": "integer"
                    },
                    "default_thread_rate_limit_per_user": {
                        "type": "integer"
                    },
                    "video_quality_mode": {
                        "type": "integer"
                    }
                }
            },
...
            "MessageCreateSchema": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "integer"
                    },
                    "content": {
                        "type": "string"
                    },
                    "nonce": {
                        "type": "string"
                    },
                    "channel_id": {
                        "type": "string"
                    },
                    "tts": {
                        "type": "boolean"
                    },
                    "flags": {
                        "type": "string"
                    },
                    "embeds": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Embed"
                        }
                    },
                    "embed": {
                        "$ref": "#/components/schemas/Embed"
                    },
                    "allowed_mentions": {
                        "type": "object",
                        "properties": {
                            "parse": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            },
                            "roles": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            },
                            "users": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            },
                            "replied_user": {
                                "type": "boolean"
                            }
                        },
                        "additionalProperties": false
                    },
                    "message_reference": {
                        "type": "object",
                        "properties": {
                            "message_id": {
                                "type": "string"
                            },
                            "channel_id": {
                                "type": "string"
                            },
                            "guild_id": {
                                "type": "string"
                            },
                            "fail_if_not_exists": {
                                "type": "boolean"
                            }
                        },
                        "additionalProperties": false,
                        "required": [
                            "channel_id",
                            "message_id"
                        ]
                    },
                    "payload_json": {
                        "type": "string"
                    },
                    "file": {},
                    "attachments": {
                        "description": "TODO: we should create an interface for attachments\nTODO: OpenWAAO<-->attachment-style metadata conversion",
                        "type": "array",
                        "items": {}
                    },
                    "sticker_ids": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
            "ChannelPermissionOverwriteSchema": {
                "type": "object",
                "properties": {
                    "allow": {
                        "type": "string"
                    },
                    "deny": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "type": {
                        "$ref": "#/components/schemas/ChannelPermissionOverwriteType"
                    }
                },
                "required": [
                    "allow",
                    "deny",
                    "id",
                    "type"
                ]
            },
...
            "BulkDeleteSchema": {
                "type": "object",
                "properties": {
                    "messages": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "required": [
                    "messages"
                ]
            },
...
        }
    },
...
    "paths": {
...
        "/users/@me/delete/": {
            "post": {
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "tags": [
                    "users"
                ]
            }
        },
...
        "/guilds/{guild_id}/delete/": {
            "post": {
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "parameters": [
                    {
                        "name": "guild_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "guild_id"
                    }
                ],
                "tags": [
                    "guilds"
                ]
            }
        },
...
        "/channels/{channel_id}/permissions/{overwrite_id}": {
            "put": {
                "x-permission-required": "MANAGE_ROLES",
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChannelPermissionOverwriteSchema"
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "channel_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "channel_id"
                    },
                    {
                        "name": "overwrite_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "overwrite_id"
                    }
                ],
                "tags": [
                    "channels"
                ]
            },
            "delete": {
                "x-permission-required": "MANAGE_ROLES",
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "parameters": [
                    {
                        "name": "channel_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "channel_id"
                    },
                    {
                        "name": "overwrite_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "overwrite_id"
                    }
                ],
                "tags": [
                    "channels"
                ]
            }
        },
        "/channels/{channel_id}/messages/bulk-delete/": {
            "post": {
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BulkDeleteSchema"
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "channel_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "channel_id"
                    }
                ],
                "tags": [
                    "channels"
                ]
            }
        },
...
        "/channels/{channel_id}/messages/{message_id}/": {
            "patch": {
                "x-right-required": "SEND_MESSAGES",
                "x-permission-required": "SEND_MESSAGES",
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MessageCreateSchema"
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "channel_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "channel_id"
                    },
                    {
                        "name": "message_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "message_id"
                    }
                ],
                "tags": [
                    "channels"
                ]
            },
            "put": {
                "x-right-required": "SEND_BACKDATED_EVENTS",
                "x-permission-required": "SEND_MESSAGES",
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MessageCreateSchema"
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "channel_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "channel_id"
                    },
                    {
                        "name": "message_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "message_id"
                    }
                ],
                "tags": [
                    "channels"
                ]
            },
            "get": {
                "x-permission-required": "VIEW_CHANNEL",
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "parameters": [
                    {
                        "name": "channel_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "channel_id"
                    },
                    {
                        "name": "message_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "message_id"
                    }
                ],
                "tags": [
                    "channels"
                ]
            },
            "delete": {
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "parameters": [
                    {
                        "name": "channel_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "channel_id"
                    },
                    {
                        "name": "message_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "message_id"
                    }
                ],
                "tags": [
                    "channels"
                ]
            }
        },
...
        "/channels/{channel_id}/": {
            "get": {
                "x-permission-required": "VIEW_CHANNEL",
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "parameters": [
                    {
                        "name": "channel_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "channel_id"
                    }
                ],
                "tags": [
                    "channels"
                ]
            },
            "delete": {
                "x-permission-required": "MANAGE_CHANNELS",
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "parameters": [
                    {
                        "name": "channel_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "channel_id"
                    }
                ],
                "tags": [
                    "channels"
                ]
            },
            "patch": {
                "x-permission-required": "MANAGE_CHANNELS",
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChannelModifySchema"
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "channel_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "channel_id"
                    }
                ],
                "tags": [
                    "channels"
                ]
            }
        },
...
        "/applications/{id}/delete": {
            "post": {
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "id"
                    }
                ],
                "tags": [
                    "applications"
                ]
            }
        },
...
    }
}